% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bjimpute}
\alias{bjimpute}
\title{ELMSurv bjimpute}
\usage{
bjimpute(y, cen, x, inibeta)
}
\arguments{
\item{y}{Survival time of training data.}

\item{cen}{The censoring indicator of training data.}

\item{x}{The covariates(predictor variables) of training data.}

\item{inibeta}{Initial values set for Buckley-James Imputation.}
}
\value{
Imputed survival times
}
\description{
A Buckley-James Esitmator
}
\references{
\itemize{
  \item Hong Wang et al (2017). A Survival Ensemble of Extreme Learning Machine. Applied Intelligence, in press.
 }
}
\seealso{
\code{\link{elm_surv}}
}
\author{
Hong Wang
}
