% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMSurvEN.R
\name{ELMSurvEN}
\alias{ELMSurvEN}
\title{ELMSurv ELMSurvEN}
\usage{
ELMSurvEN(x, y, testx, mtry = floor(sqrt(ncol(x))), trlength = 100,
  Regularization_coefficient = 10000, Kernel_type = "lin_kernel",
  Kernel_para = c(2, 1))
}
\arguments{
\item{x}{The covariates(predictor variables) of training data.}

\item{y}{Survival time and censored status of training data. Must be a Surv  \code{survival} object}

\item{testx}{The covariates(predictor variables) of test data.}

\item{mtry}{The number of covariates(predictor variables) used in each base ELM model. Default is the square root of the number of all avaibable covariates.}

\item{trlength}{The ensemle size (the number of base ELM survival models). Default is 100.}

\item{Regularization_coefficient}{Ridge or Tikhonov regularization parameter. Default is 10000. Also known as \eqn{C} in the ELM paper.}

\item{Kernel_type}{Type of kernel matrix. Currently four options avaibable. "RBF_kernel",a RBF kernel;"lin_kernel" , a linear kernel;poly_kernel ,a polynomial kernel;sigmoid_kernel, a sigmoid kernel. Default is "lin_kernel".}

\item{Kernel_para}{Parameters for different types of kernels. A single value for RBF and linear kernels. A vector for polynomial and sigmoid kernels and progam stops if only a single value is supplied. However, if the vector of values is supplied in the cases of RBF and liner kernels, only the first value will be used. Default is a vector value "c(2,1)"}
}
\value{
Object of class \code{ELMSurvEN} with elements
  \tabular{ll}{
      \code{elmsurvfit}    \tab  A list of base models \code{elm_surv} of size \code{trlength}. To retrieve a particular base model: use  elmsurvfit[[i]], where i takes values between 1 and \code{trlength} \cr
      \code{precitedtime} \tab Esitmated survival times of test data. \cr
  }
}
\description{
An Survival Ensemble of Extreme Learning Machine
}
\examples{
\dontrun{
set.seed(123)
require(ELMSurv)
require(survival)
## Survival Ensemble of ELM  with default settings
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
elmsurvfit=ELMSurvEN(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]),testx=teset[,-c(rii)])
# Get the 5th base model
basemodel=elmsurvfit[[1]]
#Print the c-index values
#library(survcomp)
#ci_elm=concordance.index(-rowMeans(elmsurvfit$precitedtime),teset$days,teset$status)[1] 
#print(ci_elm)
}
}
\references{
\itemize{
  \item Hong Wang et al (2017). A Survival Ensemble of Extreme Learning Machine. Applied Intelligence, in press.
 }
}
\seealso{
\code{\link{elm_surv}}
}
\author{
Hong Wang
}
