% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernmat}
\alias{kernmat}
\title{ELMSurv kernmat}
\usage{
kernmat(xtrain, kernel_type, kernel_para, xtest)
}
\arguments{
\item{xtrain}{The covariates(predictor variables) of training data.}

\item{kernel_type}{Type of kernel matrix. kerneltype=1,a RBF kernel;kerneltype=2 , a linear kernel;kerneltype=3 ,a polynomial kernel;kerneltype=4, a sigmoid kernel.}

\item{kernel_para}{Parameters for different types of kernels. A single value for kerneltype=1 or 2. A vector for kerneltype=3 or 4.}

\item{xtest}{The covariates(predictor variables) of test data.}
}
\value{
A kernel matrix
}
\description{
A Kermatrix Compuation
}
\references{
\itemize{
  \item Hong Wang et al (2017). A Survival Ensemble of Extreme Learning Machine. Applied Intelligence, in press.
 }
}
\seealso{
\code{\link{elm_surv}}
}
\author{
Hong Wang
}
